Renamed NavigateClient to NavigationClient
And same for NavigationServer
